+++ test/test_sheets.py
@@ -1,6 +1,8 @@
 import os
 import pytest
+from fastapi import status
 from fastapi.testclient import TestClient
+from endpoint.authentication import AccessScope
 
 from main import app
 from file_path import FilePath
@@ -42,14 +44,15 @@ def newSheet():
             "image_name": image2
         },
     ]
-    response = client.post(
-        '/sheets/?sheet_type=456090&layout=layout_1', json=body)
+    response = client.post('/sheets/?sheet_type=456090&layout=layout_1',
+                           json=body,
+                           headers={'access_token': '789conference'})
     return response
 
 
 def getSheets():
     # Get the list of images
-    response = client.get('/sheets/', headers={'access_token':'123'})
+    response = client.get('/sheets/', headers={'access_token': '123admin'})
     assert response.status_code == 200
     filenames = response.json()
     return filenames
@@ -70,21 +73,26 @@ def test_new_sheet(removeOldSheets):
 
 
 def test_get_sheet():
-    assert True
-    # TODO can we download it?
+    response = newSheet()
+    assert response.status_code == 201
 
+    filename = response.json()['filename']
+
+    response = client.get(filename, headers={'access_token': '123admin'})
+    assert response.status_code == 200
 
 def test_delete_sheet():
     response = newSheet()
     assert response.status_code == 201
     filename = response.json()['filename']
 
-    response = client.delete(filename)
+    response = client.delete(filename, headers={'access_token': '123admin'})
     assert response.status_code == 200
     assert filename == response.json()['filename']
 
     # Are they deleted
     assert not os.path.exists(filename)
+
 
 def test_wrong_layout_sheet(removeOldSheets):
     image = os.path.basename(newImage('./test/images/logo.jpg'))
@@ -108,6 +116,107 @@ def test_wrong_layout_sheet(removeOldSheets):
             "image_name": image
         },
     ]
-    response = client.post('/sheets/?sheet_type=456090&layout=invalid', json=body)
+    response = client.post('/sheets/?sheet_type=456090&layout=invalid',
+                           json=body,
+                           headers={'access_token': '789conference'})
     assert response.status_code == 400
 
+
+def test_get_sheets_access_rights():
+    response = client.get('/sheets/')
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+    response = client.get('/sheets/', headers={'access_token': '123admin'})
+    assert response.status_code == status.HTTP_200_OK
+    response = client.get('/sheets/', headers={'access_token': '456printer'})
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+    response = client.get('/sheets/', headers={'access_token': '789conference'})
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+
+
+def test_post_sheets_access_rights():
+    image1 = os.path.basename(newImage('./test/images/Kongresartikler.jpg'))
+    image2 = os.path.basename(newImage('./test/images/logo.jpg'))
+    body = [
+        {
+            "line_1": "string",
+            "line_2": "string",
+            "line_3": "string",
+            "line_4": "string",
+            "line_5": "string",
+            "qr_code": "string",
+            "image_name": image1
+        },
+        {
+            "line_1": "string",
+            "line_2": "string",
+            "line_3": "string",
+            "line_4": "string",
+            "line_5": "string",
+            "qr_code": "string",
+            "image_name": image2
+        },
+    ]
+    response = client.post('/sheets/?sheet_type=456090&layout=layout_1', json=body)
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+    response = client.post('/sheets/?sheet_type=456090&layout=layout_1', json=body, headers={'access_token': '123admin'})
+    assert response.status_code == status.HTTP_201_CREATED
+    response = client.post('/sheets/?sheet_type=456090&layout=layout_1', json=body, headers={'access_token': '789conference'})
+    assert response.status_code == status.HTTP_201_CREATED
+    response = client.post('/sheets/?sheet_type=456090&layout=layout_1', json=body, headers={'access_token': '456printer'})
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+
+def test_get_sheet_access_rights():
+    response = newSheet()
+    assert response.status_code == 201
+
+    filename = response.json()['filename']
+
+    response = client.get(filename)
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+    response = client.get(filename, headers={'access_token': '123admin'})
+    assert response.status_code == status.HTTP_200_OK
+    response = client.get(filename, headers={'access_token': '456printer'})
+    assert response.status_code == status.HTTP_200_OK
+    response = client.get(filename, headers={'access_token': '789conference'})
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+
+
+def test_delete_sheet_access_rights():
+    response = newSheet()
+    assert response.status_code == 201
+    filename = response.json()['filename']
+
+    response = client.get(filename)
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+    response = client.get(filename, headers={'access_token': '789conference'})
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+
+    response = client.get(filename, headers={'access_token': '123admin'})
+    assert response.status_code == status.HTTP_200_OK
+    
+    response = newSheet()
+    assert response.status_code == 201
+    filename = response.json()['filename']
+    response = client.get(filename, headers={'access_token': '456printer'})
+    assert response.status_code == status.HTTP_200_OK
+
+def test_access_key_checker():
+    response = client.get('/sheets/')
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+    response = client.get('/sheets/', headers={'access_token': '123admin'})
+    assert response.status_code == status.HTTP_200_OK
+    response = client.get('/sheets/', headers={'access_token': '456printer'})
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+    response = client.get('/sheets/', headers={'access_token': '789conference'})
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+    response = client.get('/sheets/', headers={'access_token': '456printer789conference'})
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+
+    response = client.get('/sheets/?access_token=12341234')
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+    response = client.get('/sheets/?access_token=123admin')
+    assert response.status_code == status.HTTP_200_OK
+    response = client.get('/sheets/?access_token=456printer')
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+    response = client.get('/sheets/?access_token=789conference')
+    assert response.status_code == status.HTTP_403_FORBIDDEN
